refactor(network): move network builder into its own module
